api:sig/verif msg sth
